+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 Release 0.0.2 (in development)
 ==============================
 
-* Added changelog
+* Added ``dfetch environment`` command.
+* Added changelog.
 
 
 Release 0.0.1 (released 2020-11-03)